+++ PeriphDriver/Source/mxc_lock.c
@@ -56,12 +56,12 @@ int mxc_get_lock(uint32_t *lock, uint32_t value)
     do {
 
         // Return if the lock is taken by a different thread
-        if(__LDREXW((volatile unsigned long *)lock) != 0) {
+        if(__LDREXW((volatile unsigned int *)lock) != 0) {
             return E_BUSY;
         }
 
         // Attempt to take the lock
-    } while(__STREXW(value, (volatile unsigned long *)lock) != 0);
+    } while(__STREXW(value, (volatile unsigned int *)lock) != 0);
 
     // Do not start any other memory access until memory barrier is complete
     __DMB();
